moved lst_align code from abscal to lstbin.py
