+++ hera_cal/__init__.py
@@ -5,5 +5,6 @@ import redcal
 import cal_formats
 import version
 import abscal
+import lstbin
 
 __version__ = version.version
